+++ frontend/src/routes/Applications/Overview.tsx
@@ -19,7 +19,7 @@ import {
 } from '../../atoms'
 import { Trans, useTranslation } from '../../lib/acm-i18next'
 import { DOC_LINKS, viewDocumentation } from '../../lib/doc-util'
-import { getAuthorizedNamespaces, rbacCreate, rbacDelete } from '../../lib/rbac-util'
+import { checkPermission, rbacCreate, rbacDelete } from '../../lib/rbac-util'
 import { queryRemoteArgoApps } from '../../lib/search'
 import { useQuery } from '../../lib/useQuery'
 import { NavigationPath } from '../../NavigationPath'
@@ -37,11 +37,6 @@ import {
     DiscoveredArgoApplicationDefinition,
     IResource,
     Subscription,
-    listProjects,
-    Namespace,
-    NamespaceApiVersion,
-    NamespaceKind,
-    ResourceAttributes,
 } from '../../resources'
 import { DeleteResourceModal, IDeleteResourceModalProps } from './components/DeleteResourceModal'
 import ResourceLabels from './components/ResourceLabels'
@@ -170,34 +165,6 @@ export const getApplicationRepos = (resource: IResource, subscriptions: Subscrip
             ]
         }
     }
-}
-
-export async function checkPermission(resourceAttributes: ResourceAttributes, setStateFn: (state: boolean) => void) {
-    // Require hub to run on OCP
-    const fetchProjects = async () => {
-        return listProjects().promise
-    }
-
-    fetchProjects().then((projects) => {
-        const namespaceArr: Namespace[] = projects.map((project) => {
-            return {
-                apiVersion: NamespaceApiVersion,
-                kind: NamespaceKind,
-                metadata: project.metadata,
-            } as Namespace
-        })
-        const fetchAuthorizedNamespaces = async () => {
-            const authorizedNamespaces = await getAuthorizedNamespaces([resourceAttributes], namespaceArr)
-            return authorizedNamespaces
-        }
-        fetchAuthorizedNamespaces().then((authorizedNamespaces) => {
-            if (authorizedNamespaces?.length > 0) {
-                setStateFn(true)
-            } else {
-                setStateFn(false)
-            }
-        })
-    })
 }
 
 export default function ApplicationsOverview() {
